feat: support onFailure setting in the parent
Today, if an expression evaluation fails, it is only possible to set a behavior at the expression level, which means that in the grouped expression, each of them needs to have its own `onFailure` attribute set.

Within this change, the `onFailure` on the parent expression is now respected, so, if the evaluation of any expression fails, the parent configuration will be used when the expression doesn't override its default throwing behavior.