feat: add support to CONTAINS operations
This change adds a new operator `CONTAINS` that can be used to assert data by checking part of its content.

When the `left` operand is a `string` it will compare if any part of the given string contains the given `right` operand.

When the `left`operand is an `array` it will compare if any elements of the given array contains the given `right` operand.