fix: Kotlin evaluator is not supporting nullable values
When evaluating an expression using the Kotlin Evaluator, when the input map contains nullable values, the evaluator throws an exception to key not found instead of returning null.

This change fixes it, returning null when a key is present but has null as value, throwing the exception only when the key is not present in the input map.